Moved ui/toolbar/utils to ui/toolbar/getitemsfromconfig.
